requirements.txt
Django==5.0.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9
djangorestframework==3.15.1
drf-yasg==1.21.7
setuptools==69.2.0
randtest==0.7
requests==2.31.0
